refactor: moved each function in controllers related to models to each its own file
